// typings/miro.d.ts
export declare namespace SDK {
// ...
  interface ITag {
    id: string;
    title: string;
    /**
     * The color of the tag
     */
    color: string | number;
    /**
     * A list of ids with the assigned tag.
     */
    widgetIds: string[];
  }
// ...
  interface IWidgetNamespaces {
    metadata: WidgetMetadata;
    capabilities: WidgetCapabilities;
    clientVisible: boolean;
  }

  /**
   * @category Widgets Manipulation
   */
  type WidgetNamespacesKeys = keyof IWidgetNamespaces;

  /**
   * Defines a basic widget and its properties
   * @category Widgets Manipulation
   */
  interface IWidget extends IWidgetNamespaces {
    readonly id: string;
    readonly type: string;
    readonly bounds: IBounds;
    readonly groupId?: string;
    readonly createdUserId: string;
    readonly lastModifiedUserId: string;
  }
// ...
  interface IBounds {
    x: number;
    y: number;
    top: number;
    left: number;
    bottom: number;
    right: number;
    width: number;
    height: number;
  }
// ...
}